+++ .gitignore
@@ -1,0 +1,2 @@
+zig-cache/
+zig-out/
