--- src/zig-cache/o/59500f8e252305058811ab24cdb88d0e/dependencies.zig
@@ -1,4 +1,0 @@
-pub const imports = struct {
-};
-pub const build_root = struct {
-};
